move ClientHello next to ResolvesServerCert
